Refactor admin tab class
