+++ Controller/Admin/TransactionTabShipping.php
@@ -19,7 +19,7 @@ class TransactionTabShipping extends TransactionTab
      *
      * @return array
      */
-    protected function _getListData(): array
+    protected function _getData(): array
     {
         return $this->_getListDataFromArray($this->oResponseMapper->getShipping());
     }
